pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "padel_league"
version = "1.0.0"
description = "Porto Padel League Flask app"
authors = [
  { name="Pedro Pacheco", email="pedropachecocontas@gmail.com" }
]
requires-python = ">=3.10"

dependencies = [
    # --- Flask core ---
    "Flask==2.3.3",
    "Werkzeug==2.3.7",
    "Jinja2==3.1.4",
    "itsdangerous==2.1.2",
    "click==8.1.7",
    "MarkupSafe==2.1.5",

    # --- DB ---
    "SQLAlchemy==1.4.52",
    "Flask-SQLAlchemy==2.5.1",
    "psycopg2-binary==2.9.9",

    # --- Migrations ---
    "Flask-Migrate==4.0.7",
    "alembic==1.13.2",

    # --- Sessions / auth / assets / mail ---
    "Flask-Session==0.6.0",
    "Flask-Login==0.6.3",
    "Flask-Assets==2.1.0",
    "Flask-Mail==0.9.1",

    # --- Google Cloud ---
    "google-cloud-storage==3.3.1",
    "google-cloud-secret-manager==2.24.0",

    # --- Misc runtime ---
    "unidecode==1.3.8",
    "typing-extensions>=4.6,<5",
    "python-dotenv>=1.0.1",
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pytest = "*"
black = "^24.0.0"
flake8 = "^7.0.0"
autoflake = "^2.3.1"
isort = "*"
mypy = "*"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    migrations
  | \.venv
  | build
  | dist
)/
'''

[tool.setuptools]
packages = ["padel_league"]

[tool.setuptools.package-dir]
padel_league = "padel_league"
